# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "TEST PRODUCT API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/product": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Add Product",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message\": \"Sucessfully Add New Product",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/product-list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get List Product",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GetListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Products"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
        "request.AddProductRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                }
            }
        },
        "request.GetListRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "keyword": {
                    "type": "string"
                }
            }
        },
        "response.Products": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                }
            }
        }
    }
}